refactor: изменит полей id c Long на Integer
Поможет измбежать лишних приведений типов

+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -13,10 +13,10 @@ import javax.validation.constraints.Size;
  * <p>Data Transfer Object для сущности "Вещь"</p>
  * <p>Dto-класс Item содержит поля:</p>
  * <ul>
- *     <li>id - уникальный идентификатор вещи, тип Long;</li>
+ *     <li>id - уникальный идентификатор вещи, тип Integer;</li>
  *     <li>name - краткое название, тип String;</li>
  *     <li>description - развёрнутое описание, тип String;</li>
- *     <li>ownerId - уникальный идентификатор владельца вещи, тип Long</li>
+ *     <li>ownerId - уникальный идентификатор владельца вещи, тип Integer</li>
  *     <li>available - статус о доступности вещи для аренды, тип Boolean.</li>
  * </ul>
  */
@@ -25,14 +25,14 @@ import javax.validation.constraints.Size;
 @Builder
 @AllArgsConstructor
 public class ItemDto {
-    @Min(value = 1L, message = "Идентификатор должен быть положительным числом")
-    private Long id;
+    @Min(value = 1, message = "Идентификатор должен быть положительным числом")
+    private Integer id;
     @NotBlank(message = "Название не может быть пустым")
     private String name;
     @Size(max = 255, message = "Размер описания не должен превышать 255 символов")
     private String description;
     @NotNull(message = "У вещи должен быть владелец")
-    @Min(value = 1L, message = "Идентификатор должен быть положительным числом")
-    private Long ownerId;
+    @Min(value = 1, message = "Идентификатор должен быть положительным числом")
+    private Integer ownerId;
     private Boolean available;
 }
