+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -12,7 +12,7 @@ import javax.validation.constraints.NotBlank;
  * <p>Data Transfer Object для сущности "Пользователь"</p>
  * <p>Dto-класс User содержит поля:</p>
  * <ul>
- *     <li>id - уникальный идентификатор пользователя, тип Long;</li>
+ *     <li>id - уникальный идентификатор пользователя, тип Integer;</li>
  *     <li>name - имя пользователя, тип String;</li>
  *     <li>email - адрес электронной почты пользователя, тип String.</li>
  * </ul>
@@ -22,8 +22,8 @@ import javax.validation.constraints.NotBlank;
 @Builder
 @AllArgsConstructor
 public class UserDto {
-    @Min(value = 1L, message = "Идентификатор должен быть положительным числом")
-    private Long id;
+    @Min(value = 1, message = "Идентификатор должен быть положительным числом")
+    private Integer id;
     private String name;
     @NotBlank
     @Email(message = "Некорректный email")
